Logging errors into database

+++ backend/springproject/src/main/java/org/springdemo/springproject/service/BookServiceImpl.java
@@ -1,58 +1,62 @@
 package org.springdemo.springproject.service;
 
-import org.springdemo.springproject.dto.BookCreateDto;
-import org.springdemo.springproject.dto.BookUpdateDto;
+import org.springdemo.springproject.dto.BookDto;
 import org.springdemo.springproject.entity.Book;
 import org.springdemo.springproject.exception.BookNotFoundException;
 import org.springdemo.springproject.repository.BookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import jakarta.validation.Valid;
+import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
-public class BookServiceImpl implements  BookService {
+public class BookServiceImpl implements BookService {
 
     @Autowired
     private BookRepository bookRepository;
 
+    @Override
     public List<Book> getAll() {
         return bookRepository.findAll();
     }
 
+    @Override
     public Book getById(Long id) {
-
         return bookRepository.findById(id)
                 .orElseThrow(() -> new BookNotFoundException("Book with ID " + id + " not found"));
     }
 
+    @Override
+    public Book createBook(BookDto bookDto) {
+        // No need for custom validation, annotations will handle it
+        Book newBook = new Book();
+        newBook.setTitle(bookDto.getTitle());
+        newBook.setAuthor(bookDto.getAuthor());
+        newBook.setPublishYear(bookDto.getPublishYear());
 
-    public Book createBook(BookCreateDto book) {
-        Book newbook = new Book();
-        newbook.setTitle(book.getTitle());
-        newbook.setAuthor(book.getAuthor());
-        newbook.setPublishYear(book.getPublishYear());
-        return bookRepository.save(newbook);
+        return bookRepository.save(newBook);
     }
 
-    public Book updateBook(Long id, BookUpdateDto book) {
-        Optional<Book> optionalBook = bookRepository.findById(id);
-        if (optionalBook.isPresent()) {
-            Book updatedbook = optionalBook.get();
-            updatedbook.setTitle(book.getTitle());
-            updatedbook.setAuthor(book.getAuthor());
-            updatedbook.setPublishYear(book.getPublishYear());
-            return bookRepository.save(updatedbook);
-        }
-        return null;
+    @Override
+    public Book updateBook(Long id,BookDto bookDto) {
+        Book existingBook = bookRepository.findById(id)
+                .orElseThrow(() -> new BookNotFoundException("Book with ID " + id + " not found cant do update"));
+
+        existingBook.setTitle(bookDto.getTitle());
+        existingBook.setAuthor(bookDto.getAuthor());
+        existingBook.setPublishYear(bookDto.getPublishYear());
+
+        return bookRepository.save(existingBook);
     }
 
+    @Override
     public void deleteBook(Long id) {
+        if (!bookRepository.existsById(id)) {
+            throw new BookNotFoundException("Book with ID " + id + " is not found, cannot delete");
+        }
+
         bookRepository.deleteById(id);
     }
-
-
-
 }
-
